- mandates loginProvider - removes loginModal
